Restore luck give functionality, add a record to the giver as well as the receiver
