+++ plugins/luckgive/commands/luck_give_cmd.rb
@@ -6,9 +6,9 @@ module AresMUSH
       attr_accessor :target_name, :reason
 
       def parse_args
-       args = cmd.parse_args(ArgParser.arg1_equals_arg2)
-       self.target_name = args.arg1
-       self.reason = args.arg2
+        args = cmd.parse_args(ArgParser.arg1_equals_arg2)
+        self.target = Character.named(args.arg1)
+        self.reason = args.arg2
       end
 
       def check_errors
@@ -25,14 +25,16 @@ module AresMUSH
 
       def handle
         date = Time.now.strftime("%Y-%m-%d")
-        ClassTargetFinder.with_a_char(self.target_name) do |model|
-          LuckRecord.create(date: date, character: model, reason: self.reason, from: enactor.name)
-          client.emit_success t('luckgive.given_luck', :name => model.name)
-          message = t('luckgive.received_luck', :from => enactor.name)
-          Login.emit_if_logged_in model, message
-          FS3Skills.modify_luck(model, 1)
-          FS3Skills.modify_luck(enactor, -1)
-        end
+        LuckRecord.create(date: date, character: self.target, reason: self.reason, from: enactor.name, to: self.target.name)
+        LuckRecord.create(date: date, character: enactor, reason: self.reason, from: enactor.name, to: self.target.name)
+
+        client.emit_success t('luckgive.given_luck', :name => self.target.name)
+        message = t('luckgive.received_luck', :from => enactor.name)
+        Login.emit_if_logged_in self.target, message
+        Login.notify(self.target, :luck, message, nil)
+
+        FS3Skills.modify_luck(self.target, 1)
+        FS3Skills.modify_luck(enactor, -1)
       end
 
     end
