+++ plugins/luckgive/locales/locale_en.yml
@@ -6,4 +6,4 @@ en:
     max_luck: "%{name} already has the maximum amount of luck."
     given_luck: "You have given %{name} a luck point."
     received_luck: "%{from} has given you a luck point."
-    luck_record_title: "LUCK GIVEN TO %{name}"
+    luck_record_title: "LUCK TRANSACTIONS FOR %{name}"
