+++ plugins/luckgive/public/luck_give_char.rb
@@ -1,0 +1,5 @@
+module AresMUSH
+  class Character < Ohm::Model
+    collection :luck_record, "AresMUSH::LuckRecord"
+  end
+end
