+++ plugins/luckgive/public/luck_record.rb
@@ -1,0 +1,12 @@
+module AresMUSH
+  class LuckRecord < Ohm::Model
+    include ObjectModel
+
+    index :character
+    reference :character, "AresMUSH::Character"
+    attribute :reason
+    attribute :from
+    attribute :to
+
+  end
+end
